milestone

+++ sign/sign.c
@@ -10,7 +10,7 @@
 #define FIELDS_OFFSET 0x110U
 const uint32_t pattern[4] = {0xAAAAAAAAU, 0xBBBBBBBBU, 0xCCCCCCCCU, 0xDDDDDDDDU};
 
-static unsigned long sw_crc32_table[256] = {
+static uint32_t sw_crc32_table[256] = {
 	0x00000000, 0x04c11db7, 0x09823b6e, 0x0d4326d9, // 0..3
 	0x130476dc, 0x17c56b6b, 0x1a864db2, 0x1e475005, // 4..7
 	0x2608edb8, 0x22c9f00f, 0x2f8ad6d6, 0x2b4bcb61, // 8..B
